fix: rename directory to rename file
