+++ components/Card2/CardContent.tsx
