+++ components/Card2/CardFooter.test.tsx
