+++ components/Card2/CardFooter.tsx
